+++ resources/lang/es/blog.php
@@ -10,5 +10,6 @@ return [
     'title' => 'Título de la aventura',
     'content' => 'Contenido',
     'content_placeholder' => 'Escribe historias maravillosas <i class="fad fa-smile-wink colored"></i>',
+    'class_blog' => 'Esta aventura es compartida con toda la clase, ¡lo que escribas será visible por todos!',
 
 ];